Revert "research optimal lambda - 2 - auto lambda for bm25 - debug 2"
This reverts commit bd771c6344d70ed5c7e20bef0316758a5e8854d8.

+++ src/main/java/experiments/TRECDLQPPEvaluator.java
@@ -142,7 +142,7 @@ public class TRECDLQPPEvaluator {
 
             TauAndSARE tauAndSARE = runExperiment(baseModelName,
                         searcher, knnRelModel, evaluatorTrain,
-                        trainQueries, topDocsMap, (float)0.5, numVariants, targetMetric, true); // lambda here has no meaning because of the auto setting
+                        trainQueries, topDocsMap, 0.5, numVariants, targetMetric, true); // lambda here has no meaning because of the auto setting
             System.out.println(String.format("Train on %s -- (%s, %d): tau = %.4f",
                         trainQueryFile, "-auto-", numVariants, tauAndSARE.tau, tauAndSARE.sare)); 
         }
